Minor optimization; make flycheck-projectile-error-list-mode private
flycheck-projectile--handle-flycheck-off now always reloads project errors and
does not check if the current buffer is a project buffer. That is okay, because
it is only ever called when flycheck-projectile--project-buffer-mode is on too,
so only for project buffers. As such, the check was redundant.

flycheck-projectile-error-list-mode did not have -- name, because flycheck's
didn't either. However, I decided that it is an implementation detail and as
such should not be exposed.