Acquire flycheck's tabulated list formats at run-time
No longer wrap `flycheck-error-list-*' in (with-eval-after-load). The mode in
which that is done is private anyway and called only after `flycheck' is already
loaded.

Add a single-test ert test-suite asserting that `flycheck-projectile' lazy-loads
modules correctly and depend on `ert-runner' (development) for that.

No longer use Elsa, since it didn't really ever work.