Fix wrong type argument stringp, nil errors
In DOOM Emacs, buffers that have no default directory and no
buffer-file-name cannot be shown without errors for some reason. This
was the case the case for the error list buffer, because the
default-directory was set to nil for reasons that are now unneccessary.
Don't do that, meaning that there are no more errors.