Don't lazy-load `flycheck'
Lazy-loading flycheck and `eval-when-compile' evaluating its mode-map
prevents `flycheck-projectile-error-list-mode' from acquiring
keybindings defined for `flycheck-error-list-mode' in an
`eval-after-load' block. Also, it is unnecessary, given that the
autoloaded main entry-point of this package,
`flycheck-projectile-list-errors', would eventually cause it to be
loaded anyway.

Change the API of some functions to no longer state that flycheck needs
to already be loaded for them to work; because of this, bump the
version.